requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dateparser==1.1.1
docopt==0.6.2
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
pandas==1.4.2
passlib==1.7.4
pipreqs==0.4.11
pprintpp==0.4.0
pycparser==2.21
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
typing==3.7.4.3
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
urllib3==1.26.9
websockets==10.3
Werkzeug==2.1.1
WTForms==3.0.1
yarg==0.1.9
yarl==1.7.2
zipp==3.8.0
